% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatchers.R
\name{CovarEstim}
\alias{CovarEstim}
\title{Covariance matrix estimator}
\usage{
CovarEstim(x, type, ...)
}
\arguments{
\item{x}{a matrix of asset returns. Rows represent different assets, columns- observations.}

\item{type}{a character. The the estimation method to use.}

\item{...}{arguments to pass to estimators}
}
\description{
Function dispatcher for covariance estimation.
}
\examples{
n<-3e2 # number of realizations
p<-.5*n # number of assets

x <- matrix(data = rnorm(n*p), nrow = p, ncol = n)

Mtrx_naive <- CovarEstim(x, type="naive")

TM <- matrix(0, p, p)
diag(TM) <- 1
Mtrx_bgp <- CovarEstim(x, type="BGP14", TM=TM, SCM=Mtrx_naive)

Mtrx_bgp <- CovarEstim(x, type="LW20")#, TM=TM)
}
