% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_shrink.R
\name{CovShrinkBGP14}
\alias{CovShrinkBGP14}
\title{For shrinkage of the covariance matrix (2014, JMVA, Strong convergence)}
\usage{
CovShrinkBGP14(n, TM, SCM)
}
\arguments{
\item{n}{the number of observations of x.}

\item{TM}{a shrinkage target.}

\item{SCM}{the sample covariance matrix.}
}
\description{
It returns the estimator CovShrink(TM,SCM)=alfa*SCM+beta*TM,
where SCM should be the sample covariance matrix and TM is a shrinkage target.
In most cases, we took TM=I but could be any deterministic pos. def. matrix.
In order to use the estimator for portfolio weights one needs to invert it,
i.e., solve(CovShrink(TM,SCM)).
}
\references{
\insertRef{BGP2014}{hdsp}
}
