% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatchers.R
\name{EUShrinkPortfolio}
\alias{EUShrinkPortfolio}
\title{Shrinkage portfolio}
\usage{
EUShrinkPortfolio(x, gamma, type, subtype, ...)
}
\arguments{
\item{x}{a matrix of asset returns. Rows represent different assets, columns- observations.}

\item{gamma}{a numerical variable. Investors attitude towards risk.}

\item{type}{a character. The type of methods to use to construct the portfolio.}

\item{subtype}{a character. The exact method to use within the type.}

\item{...}{arguments to pass to portfolio constructors}
}
\description{
Function dispatcher for portfolio construction
}
\examples{
n<-3e2 # number of realizations
p<-.5*n # number of assets
b<-rep(1/p,p)
gamma<-1

x <- matrix(data = rnorm(n*p), nrow = p, ncol = n)

test <- EUShrinkPortfolio(x=x, gamma=gamma, type='weights', b=b)
str(test)

test <- EUShrinkPortfolio(x=x, gamma=gamma, type='mean', subtype='Bayes-Stein')
str(test)

test <- EUShrinkPortfolio(x=x, gamma=gamma, type='mean', subtype='James-Stein')
str(test)
}
