% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_custom_portfol.R
\name{ExUtil_portfolio_custom}
\alias{ExUtil_portfolio_custom}
\title{A helper function for ExUtil_portfolio}
\usage{
ExUtil_portfolio_custom(mean_vec, inv_cov_mtrx, gamma)
}
\arguments{
\item{mean_vec}{mean vector or list of asset returns.}

\item{inv_cov_mtrx}{the inverse covariance matrix of asset returns. Could be a matrix or a data frame.}

\item{gamma}{a numerical variable. Investors attitude towards risk.}
}
\description{
A helper function for ExUtil_portfolio
}
\examples{
n<-3e2 # number of realizations
p<-.5*n # number of assets
gamma<-1

x <- matrix(data = rnorm(n*p), nrow = p, ncol = n)

# Simple EU portfolio
cov_mtrx <- Sigma_sample_estimator(x)
invSS <- solve(cov_mtrx)
means <- rowMeans(x)

cust_port_simp <- ExUtil_portfolio_custom(mean_vec=means, inv_cov_mtrx=invSS, gamma=2)
str(cust_port_simp)
}
