% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_shrink.R
\name{InvCovShrinkBGP16}
\alias{InvCovShrinkBGP16}
\title{For the shrinkage of the inverse covariance matrix.}
\usage{
InvCovShrinkBGP16(n, p, TM, iSCM)
}
\arguments{
\item{n}{the number of observations}

\item{p}{the number of variables (rows of the covariance matrix)}

\item{TM}{the target matrix for shrinkage of the inverse covariance matrix}

\item{iSCM}{the inverse of the sample covariance matrix}
}
\description{
iSCM=solve(S) with S sample covariance matrix and TM is again a target matrix,
for example TM=I. Thus, InvCovShrink(solve(S), TM) will return the estimator of
the inverse covariance matrix (no need to invert anymore). n- the number of observations of X.
}
\references{
\insertRef{BGP2016}{hdsp}
}
