% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mean_shrink.R
\name{mean_bs}
\alias{mean_bs}
\title{Bayes-Stein shrinkage mean estimator}
\usage{
mean_bs(x, mu_0)
}
\arguments{
\item{x}{a numeric matrix. Rows represent different variables, columns- observations.}

\item{mu_0}{a numeric value. The scaling of the target for shrinkage of the mean vector.}
}
\description{
Bayes-Stein shrinkage mean estimator
}
\references{
\insertRef{Jorion1986}{hdsp}
}
