% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatchers.R
\name{MeanEstim}
\alias{MeanEstim}
\title{Mean vector shrinkage estimator}
\usage{
MeanEstim(x, type, ...)
}
\arguments{
\item{x}{a matrix of asset returns. Rows represent different assets, columns- observations.}

\item{type}{a character. The the estimation method to use.}

\item{...}{arguments to pass to estimators}
}
\value{
a numeric vector
}
\description{
Essentially it is a function dispatcher for mean value estimators that chooses a method according to the type argument.
}
\details{
The available estimation methods are:\tabular{lll}{
   Function \tab Paper \tab Type \cr
   .rowMeans \tab  \tab naive \cr
   \code{\link{mean_bs}} \tab Jorion 1986 \tab bs \cr
   \code{\link{mean_js}} \tab Jorion 1986 \tab js \cr
   \code{\link{mean_bop19}} \tab Bodnar et al 2019 \tab BOP19 \cr
}
}
\examples{
n<-3e2 # number of realizations
p<-.5*n # number of assets

x <- matrix(data = rnorm(n*p), nrow = p, ncol = n)

Mean_naive <- MeanEstim(x, type="naive")

mu_0 <- rep(1/p, p)
Mean_BOP <- MeanEstim(x, type="BOP19", mu_0=mu_0)

}
