% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_cov_portfol.R
\name{new_ExUtil_portfolio_cov_BGP14}
\alias{new_ExUtil_portfolio_cov_BGP14}
\title{Bodnar, Gupta, Parolya 2014}
\usage{
new_ExUtil_portfolio_cov_BGP14(x, gamma, TM)
}
\arguments{
\item{x}{a matrix of asset returns. Rows represent different assets, columns- observations.}

\item{gamma}{a numerical variable. Investors attitude towards risk.}

\item{TM}{the target matrix for shrinkage of covariance}
}
\description{
Bodnar, Gupta, Parolya 2014
}
