% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_custom_portfol.R
\name{new_ExUtil_portfolio_custom}
\alias{new_ExUtil_portfolio_custom}
\title{Custom EU portfolio constructors}
\usage{
new_ExUtil_portfolio_custom(mean_vec, inv_cov_mtrx, gamma)
}
\arguments{
\item{mean_vec}{mean vector of asset returns}

\item{inv_cov_mtrx}{the inverse covariance matrix of asset returns}

\item{gamma}{a numerical variable. Investors attitude towards risk.}
}
\description{
Custom EU portfolio constructors
}
\examples{
n<-3e2 # number of realizations
p<-.5*n # number of assets
gamma<-1

x <- matrix(data = rnorm(n*p), nrow = p, ncol = n)

# Simple EU portfolio
cov_mtrx <- Sigma_sample_estimator(x)
invSS <- solve(cov_mtrx)
means <- rowMeans(x)

cust_port_simp <- new_ExUtil_portfolio_custom(mean_vec=means, inv_cov_mtrx=invSS, gamma=2)
str(cust_port_simp)

# Portfolio with Bayes-Stein shrinked means
# and a Ledoit and Wolf estimator for covariance matrix
TM <- matrix(0, p, p)
diag(TM) <- 1
cov_mtrx <- CovarEstim(x, type="LW20", TM=TM)
invSS <- solve(cov_mtrx)
means <- rowMeans(x)

cust_port_BS_LW <- new_ExUtil_portfolio_custom(mean_vec=means, inv_cov_mtrx=invSS, gamma=2)
str(cust_port_BS_LW)
}
