% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_mean_portfol.R
\name{new_ExUtil_portfolio_mean_BOP}
\alias{new_ExUtil_portfolio_mean_BOP}
\title{constructor of EU portfolio object. Type=mean, Mean.type=BOP.}
\usage{
new_ExUtil_portfolio_mean_BOP(x, gamma, mu_0 = mu_0)
}
\arguments{
\item{x}{a matrix of asset returns. Rows represent different assets, columns- observations.}

\item{gamma}{a numerical variable. Investors attitude towards risk.}

\item{mu_0}{a numeric vector. The target for shrinkage of the mean vector of asset returns.}
}
\description{
constructor of EU portfolio object. Type=mean, Mean.type=BOP.
}
\examples{
n<-3e2 # number of realizations
p<-.5*n # number of assets
gamma<-1
mu_0 <- rep(1,p)

x <- matrix(data = rnorm(n*p), nrow = p, ncol = n)

test <- new_ExUtil_portfolio_mean_BOP(x=x, gamma=gamma, mu_0=mu_0)
str(test)
}
