% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_prototypes.R
\name{new_ExUtil_portfolio_mean_JamesStein}
\alias{new_ExUtil_portfolio_mean_JamesStein}
\title{constructor of EU portfolio object. Type=mean, Mean.type=James-Stein.}
\usage{
new_ExUtil_portfolio_mean_JamesStein(x, gamma, mu_0 = 0)
}
\description{
constructor of EU portfolio object. Type=mean, Mean.type=James-Stein.
}
\examples{
n<-3e2 # number of realizations
p<-.5*n # number of assets
gamma<-1

x <- matrix(data = rnorm(n*p), nrow = p, ncol = n)

test <- new_ExUtil_portfolio_mean_JamesStein(x=x, gamma=gamma, mu_0=0)
str(test)
}
