% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_naive_portfol.R
\name{new_ExUtil_portfolio_naive}
\alias{new_ExUtil_portfolio_naive}
\title{constructor of EU portfolio object. Type=naive.}
\usage{
new_ExUtil_portfolio_naive(x, gamma)
}
\arguments{
\item{x}{a matrix or a data frame of asset returns. Rows represent different assets, columns- observations.}

\item{gamma}{a numeric variable. Investors attitude towards risk.}
}
\value{
an object of class ExUtil_portfolio\tabular{ll}{
   Element \tab Description \cr
   call \tab the function call with which it was created \cr
   cov_mtrx \tab the sample covariance matrix of the assets \cr
   inv_cov_mtrx \tab the inverse of the sample covariance matrix \cr
   means \tab sample mean vector estimate for the assets \cr
   W_EU_hat \tab portfolio weights_sample estimate \cr
   Port_Var \tab portfolio variance \cr
   Port_mean_return \tab portfolio mean returns \cr
   Sharpe \tab portfolio Sharpe coefficient \cr
}
}
\description{
constructor of EU portfolio object. Type=naive.
}
\examples{
n<-3e2 # number of realizations
p<-.5*n # number of assets
gamma<-1

x <- matrix(data = rnorm(n*p), nrow = p, ncol = n)

test <- new_ExUtil_portfolio_naive(x=x, gamma=gamma)
str(test)
}
