% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_weights_portfol.R
\name{new_ExUtil_portfolio_weights_BDOPS20}
\alias{new_ExUtil_portfolio_weights_BDOPS20}
\title{constructor of EU portfolio object. IEEE 2020}
\usage{
new_ExUtil_portfolio_weights_BDOPS20(x, gamma, b, alph)
}
\arguments{
\item{x}{a matrix or a data frame of asset returns. Rows represent different assets, columns- observations.}

\item{gamma}{a numeric variable. Investors attitude towards risk.}

\item{b}{a numeric variable. The target for weight shrinkage.}

\item{alph}{a numeric variable. The level of confidence for weight intervals}
}
\description{
constructor of EU portfolio object. IEEE 2020
}
\examples{
# Assets with a diagonal covariance matrix

n<-3e2 # number of realizations
p<-.5*n # number of assets
b<-rep(1/p,p)
gamma<-1

x <- matrix(data = rnorm(n*p), nrow = p, ncol = n)

test <- new_ExUtil_portfolio_weights_BDOPS20(x=x, gamma=gamma, b=b, alph=0.05)
str(test)

# Assets with a non-diagonal covariance matrix

Mtrx <- RandCovMtrx(n=n, p=p, q=20.55, mu=seq(0.2,-0.2, length.out=p))
x <- t(MASS::mvrnorm(n=n , mu=rep(0,p), Sigma=Mtrx))

test <- new_ExUtil_portfolio_weights_BDOPS20(x=x, gamma=gamma, b=b, alph=0.05)
str(test)
}
\references{
\insertRef{BDOPS2020}{hdsp}
}
