% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_weights_portfol.R
\name{new_GMV_portfolio_weights_BDPS19}
\alias{new_GMV_portfolio_weights_BDPS19}
\title{constructor of GMVA portfolio object.}
\usage{
new_GMV_portfolio_weights_BDPS19(x, b, alph)
}
\arguments{
\item{x}{a matrix or a data frame of asset returns. Rows represent different assets, columns- observations.}

\item{b}{a numeric value. The target for weight shrinkage.}

\item{alph}{a numeric value. The level of confidence for weight intervals}
}
\description{
constructor of GMVA portfolio object.
}
\examples{
library(MASS)
n<-3e2 # number of realizations
p<-.5*n # number of assets
b<-rep(1/p,p)

Mtrx <- RandCovMtrx(n=n, p=p, q=20.55, mu=seq(0.2,-0.2, length.out=p))
x <- mvrnorm(n=n , mu=rep(0,p), Sigma=Mtrx)

test <- new_GMV_portfolio_weights_BDPS19(x=x, b=b, alph=0.05)
str(test)

}
\references{
\insertRef{BDOPS2020}{hdsp}
}
