% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_shrink.R
\name{nonlin_shrinkLW}
\alias{nonlin_shrinkLW}
\title{Ledoit-Wolf Nonlinear shrinkage estimator (Annals of Statistics (2020))}
\usage{
nonlin_shrinkLW(x)
}
\arguments{
\item{x}{a numeric matrix. Rows represent different variables, columns- observations.}
}
\description{
Returns the estimator of the covariance matrix.
}
\references{
\insertRef{LW2020}{hdsp}
}
