% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimators.R
\name{Q_hat_n}
\alias{Q_hat_n}
\title{Q_n_hat}
\usage{
Q_hat_n(x)
}
\arguments{
\item{x}{numeric matrix in which columns are independent realizations of asset returns}
}
\description{
Q estimator from (6) of EU_IEEE.
}
