% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDOPS_2020.R
\name{Sigma_sample_estimator}
\alias{Sigma_sample_estimator}
\title{Simple sample covariance estimator}
\usage{
Sigma_sample_estimator(x)
}
\arguments{
\item{x}{a numeric matrix or a data frame in which columns are independent realizations of asset returns}
}
\description{
It computes covariance of matrix x as follows:
\deqn{\hat\Sigma_n = \frac{1}{n-1} \sum_{j=1}^n (x_j - \bar x_n)(x_j - \bar x_n)^T
,\quad \bar x_n = \frac{1}{n} \sum_{j=1}^n x_j}
}
\examples{
p<-5 # number of assets
n<-1e1 # number of realizations

x <-matrix(data = rnorm(n*p), nrow = p, ncol = n)
Sigma_sample_estimator(x)
}
\references{
\insertRef{BDOPS2020}{hdsp}
}
