% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_custom_portfol.R
\name{validate_ExUtil_portfolio}
\alias{validate_ExUtil_portfolio}
\title{A validator for ExUtil_portfolio}
\usage{
validate_ExUtil_portfolio(x)
}
\arguments{
\item{x}{Object of class ExUtil_portfolio.}
}
\description{
A validator for ExUtil_portfolio
}
\examples{
n<-3e2 # number of realizations
p<-.5*n # number of assets
gamma<-1

x <- matrix(data = rnorm(n*p), nrow = p, ncol = n)

# Simple EU portfolio
cov_mtrx <- Sigma_sample_estimator(x)
invSS <- solve(cov_mtrx)
means <- rowMeans(x)

cust_port_simp <- new_ExUtil_portfolio_custom(mean_vec=means, inv_cov_mtrx=invSS, gamma=2)
str(validate_ExUtil_portfolio(cust_port_simp))
}
