% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDOPS_2020.R
\name{W_EU_hat}
\alias{W_EU_hat}
\title{W_EU estimator}
\usage{
W_EU_hat(x, gamma)
}
\arguments{
\item{x}{numeric matrix in which columns are independent realizations of asset returns}

\item{gamma}{positive numeric. Investors attitude towards risk}
}
\description{
EU portfolio weights estimator from (6) of EU_IEEE.
}
