% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatchers.R
\name{CovarEstim}
\alias{CovarEstim}
\title{Covariance matrix estimator}
\usage{
CovarEstim(x, type, ...)
}
\arguments{
\item{x}{a matrix of asset returns. Rows represent different assets, columns- observations.}

\item{type}{a character. The the estimation method to use.}

\item{...}{arguments to pass to estimators}
}
\description{
Essentially it is a function dispatcher for covariance estimation that chooses a method according to the type argument.
}
\details{
The available estimation methods are:\tabular{lll}{
   Function \tab Paper \tab Type \cr
   \code{\link{Sigma_sample_estimator}} \tab  \tab naive \cr
   \code{\link{CovShrinkBGP14}} \tab Bodnar et al 2014 \tab BGP14 \cr
   \code{\link{nonlin_shrinkLW}} \tab Ledoit & Wolf 2020 \tab LW20 \cr
}
}
\examples{
n<-3e2 # number of realizations
p<-.5*n # number of assets

x <- matrix(data = rnorm(n*p), nrow = p, ncol = n)

Mtrx_naive <- CovarEstim(x, type="naive")

TM <- matrix(0, p, p)
diag(TM) <- 1
Mtrx_bgp <- CovarEstim(x, type="BGP14", TM=TM, SCM=Mtrx_naive)

Mtrx_bgp <- CovarEstim(x, type="LW20")
}
