% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatchers.R
\name{EUShrinkPortfolio}
\alias{EUShrinkPortfolio}
\title{Shrinkage Expected Utility portfolio}
\usage{
EUShrinkPortfolio(x, gamma, type, subtype, ...)
}
\arguments{
\item{x}{a matrix or a data frame of asset returns. Rows represent different assets, columns- observations.}

\item{gamma}{a numerical variable. Investors attitude towards risk.}

\item{type}{a character. The type of methods to use to construct the portfolio.}

\item{subtype}{a character. The exact method to use within the type.}

\item{...}{arguments to pass to portfolio constructors}
}
\value{
an object of class ExUtil_portfolio potentially with a subclass.
See \code{\link{new_ExUtil_portfolio_custom}} for the details of the class.
}
\description{
The main function for EU portfolio construction. It is a dispatcher using methods according
to arguments type and subtype..
}
\details{
The available estimation methods are:\tabular{lll}{
   Function \tab Paper \tab Type \cr
   \code{\link{new_ExUtil_portfolio_mean_BayesStein}} \tab Jorion 1986 \tab mean \cr
   \code{\link{new_ExUtil_portfolio_mean_JamesStein}} \tab Jorion 1986 \tab mean \cr
   \code{\link{new_ExUtil_portfolio_mean_BOP19}} \tab Bodnar et al 2019 \tab mean \cr
   \code{\link{new_ExUtil_portfolio_cov_LW02}} \tab Ledoit & Wolf 2020 \tab cov \cr
   \code{\link{new_ExUtil_portfolio_cov_BGP14}} \tab Bodnar et al 2014 \tab cov \cr
   \code{\link{new_ExUtil_portfolio_icov_BGP16}} \tab Bodnar et al 2016 \tab inv_cov \cr
   \code{\link{new_ExUtil_portfolio_weights_BDOPS20}} \tab Bodnar et al 2020 \tab weights \cr
}
}
\examples{
n<-3e2 # number of realizations
p<-.5*n # number of assets
b<-rep(1/p,p)
gamma<-1

x <- matrix(data = rnorm(n*p), nrow = p, ncol = n)

test <- EUShrinkPortfolio(x=x, gamma=gamma, type='weights', b=b)
str(test)

test <- EUShrinkPortfolio(x=x, gamma=gamma, type='mean', subtype='Bayes-Stein')
str(test)

test <- EUShrinkPortfolio(x=x, gamma=gamma, type='mean', subtype='James-Stein')
str(test)
}
