% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_cov_portfol.R
\name{new_ExUtil_portfolio_cov_LW02}
\alias{new_ExUtil_portfolio_cov_LW02}
\title{Ledoit and Wolf 2020}
\usage{
new_ExUtil_portfolio_cov_LW02(x, gamma)
}
\arguments{
\item{x}{a matrix of asset returns. Rows represent different assets, columns- observations.}

\item{gamma}{a numerical variable. Investors attitude towards risk.}
}
\description{
Ledoit and Wolf 2020
}
