% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_custom_portfol.R
\name{new_ExUtil_portfolio_custom}
\alias{new_ExUtil_portfolio_custom}
\title{A constructor for class ExUtil_portfolio}
\usage{
new_ExUtil_portfolio_custom(mean_vec, inv_cov_mtrx, gamma)
}
\arguments{
\item{mean_vec}{mean vector of asset returns}

\item{inv_cov_mtrx}{the inverse covariance matrix of asset returns}

\item{gamma}{a numerical variable. Investors attitude towards risk.}
}
\value{
Object of S3 class ExUtil_portfolio.

This class is designed to describe Expected Utility portfolios. It comprises portfolio weights
W_EU_hat, mean vector means and an inverse covariance inv_cov_mtrx. W_EU_hat and means both
have the form of a numeric vector, while inv_cov_mtrx is a matrix. The direct covariance matrix
is not included in order to cover cases where only the inverse one exists.
}
\description{
A light-weight constructor of objects of S3 class ExUtil_portfolio. This  function is for development purposes.
A helper function equipped with error messages and allowing more flexible input is \code{\link{ExUtil_portfolio_custom}}.
}
\examples{
n<-3e2 # number of realizations
p<-.5*n # number of assets
gamma<-1

x <- matrix(data = rnorm(n*p), nrow = p, ncol = n)

# Simple EU portfolio
cov_mtrx <- Sigma_sample_estimator(x)
invSS <- solve(cov_mtrx)
means <- rowMeans(x)

cust_port_simp <- new_ExUtil_portfolio_custom(mean_vec=means, inv_cov_mtrx=invSS, gamma=2)
str(cust_port_simp)

# Portfolio with Bayes-Stein shrinked means
# and a Ledoit and Wolf estimator for covariance matrix
TM <- matrix(0, p, p)
diag(TM) <- 1
cov_mtrx <- CovarEstim(x, type="LW20", TM=TM)
invSS <- solve(cov_mtrx)
means <- rowMeans(x)

cust_port_BS_LW <- new_ExUtil_portfolio_custom(mean_vec=means, inv_cov_mtrx=invSS, gamma=2)
str(cust_port_BS_LW)
}
