% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_cov_portfol.R
\name{new_ExUtil_portfolio_icov_BGP16}
\alias{new_ExUtil_portfolio_icov_BGP16}
\title{Portfolio with inverse covariance shrinkage}
\usage{
new_ExUtil_portfolio_icov_BGP16(x, gamma, TM)
}
\arguments{
\item{x}{a matrix or a data frame of asset returns. Rows represent different assets, columns- observations.}

\item{gamma}{a numerical variable. Investors attitude towards risk.}

\item{TM}{the target matrix for shrinkage of the inverse covariance matrix}
}
\description{
Portfolio with inverse covariance shrinkage
}
\references{
\insertRef{BGP2016}{hdsp}
}
