% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_mean_portfol.R
\name{new_ExUtil_portfolio_mean_BayesStein}
\alias{new_ExUtil_portfolio_mean_BayesStein}
\title{constructor of EU portfolio object. Type=mean, Mean.type=Bayes-Stein.}
\usage{
new_ExUtil_portfolio_mean_BayesStein(x, gamma, mu_0 = 0)
}
\arguments{
\item{x}{a matrix of asset returns. Rows represent different assets, columns- observations.}

\item{gamma}{a numerical variable. Investors attitude towards risk.}

\item{mu_0}{a numeric value. The scaling of the target for shrinkage of the mean vector.}
}
\description{
constructor of EU portfolio object. Type=mean, Mean.type=Bayes-Stein.
}
\examples{
n<-3e2 # number of realizations
p<-.5*n # number of assets
gamma<-1

x <- matrix(data = rnorm(n*p), nrow = p, ncol = n)

test <- new_ExUtil_portfolio_mean_BayesStein(x=x, gamma=gamma, mu_0 = 0)
str(test)
}
