% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_naive_portfol.R
\name{new_ExUtil_portfolio_naive}
\alias{new_ExUtil_portfolio_naive}
\title{constructor of EU portfolio object. Type=naive.}
\usage{
new_ExUtil_portfolio_naive(x, gamma)
}
\arguments{
\item{x}{a matrix or a data frame of asset returns. Rows represent different assets, columns- observations.}

\item{gamma}{a numeric variable. Investors attitude towards risk.}
}
\description{
constructor of EU portfolio object. Type=naive.
}
\examples{
n<-3e2 # number of realizations
p<-.5*n # number of assets
gamma<-1

x <- matrix(data = rnorm(n*p), nrow = p, ncol = n)

test <- new_ExUtil_portfolio_naive(x=x, gamma=gamma)
str(test)
}
