% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_weights_portfol.R
\name{new_ExUtil_portfolio_weights_BDOPS20}
\alias{new_ExUtil_portfolio_weights_BDOPS20}
\title{constructor of EU portfolio object. IEEE 2020}
\usage{
new_ExUtil_portfolio_weights_BDOPS20(x, gamma, b, alph)
}
\arguments{
\item{x}{a matrix or a data frame of asset returns. Rows represent different assets, columns- observations.}

\item{gamma}{a numeric variable. Investors attitude towards risk.}

\item{b}{a numeric variable. The target for weight shrinkage.}

\item{alph}{a numeric variable. The level of confidence for weight intervals}
}
\value{
an object of class ExUtil_portfolio with a subclass ... .\tabular{ll}{
   Element \tab Description \cr
   call \tab the function call with which it was created \cr
   cov_mtrx \tab the sample covariance matrix of the assets \cr
   inv_cov_mtrx \tab the inverse of the sample covariance matrix \cr
   means \tab sample mean vector estimate for the assets \cr
   W_EU_hat \tab portfolio weights_sample estimate \cr
   weights \tab shrunk portfolio weights \cr
   alpha \tab shrinkage intensity for the weights \cr
   Port_Var \tab portfolio variance \cr
   Port_mean_return \tab portfolio mean returns \cr
   Sharpe \tab portfolio Sharpe coefficient \cr
   weight_intervals \tab A data frame \cr
}
}
\description{
constructor of EU portfolio object. IEEE 2020
}
\examples{
# Assets with a diagonal covariance matrix

n<-3e2 # number of realizations
p<-.5*n # number of assets
b<-rep(1/p,p)
gamma<-1

x <- matrix(data = rnorm(n*p), nrow = p, ncol = n)

test <- new_ExUtil_portfolio_weights_BDOPS20(x=x, gamma=gamma, b=b, alph=0.05)
summary(test)

# Assets with a non-diagonal covariance matrix

Mtrx <- RandCovMtrx(n=n, p=p, q=20.55)
x <- t(MASS::mvrnorm(n=n , mu=rep(0,p), Sigma=Mtrx))

test <- new_ExUtil_portfolio_weights_BDOPS20(x=x, gamma=gamma, b=b, alph=0.05)
str(test)
}
\references{
\insertRef{BDOPS2020}{hdsp}
}
