% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_weights_portfol.R
\name{new_ExUtil_portfolio_weights_BDOPS20}
\alias{new_ExUtil_portfolio_weights_BDOPS20}
\title{constructor of EU portfolio object. IEEE 2020}
\usage{
new_ExUtil_portfolio_weights_BDOPS20(x, gamma, b)
}
\arguments{
\item{x}{a matrix or a data frame of asset returns. Rows represent different assets, columns- observations.}

\item{gamma}{a numerical variable. Investors attitude towards risk.}

\item{b}{a numeric value. The target for weight shrinkage.}
}
\description{
constructor of EU portfolio object. IEEE 2020
}
\examples{
n<-3e2 # number of realizations
p<-.5*n # number of assets
b<-rep(1/p,p)
gamma<-1

x <- matrix(data = rnorm(n*p), nrow = p, ncol = n)

test <- new_ExUtil_portfolio_weights_BDOPS20(x=x, gamma=gamma, b=b)
str(test)
}
\references{
\insertRef{BDOPS2020}{hdsp}
}
