% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_weights_portfol.R
\name{new_GMV_portfolio_weights_BDPS19}
\alias{new_GMV_portfolio_weights_BDPS19}
\title{constructor of GMVA portfolio object.}
\usage{
new_GMV_portfolio_weights_BDPS19(x, b, alph)
}
\arguments{
\item{x}{a matrix or a data frame of asset returns. Rows represent different assets, columns- observations.}

\item{b}{a numeric value. The target for weight shrinkage.}

\item{alph}{a numeric variable. The level of confidence for weight intervals}
}
\description{
constructor of GMVA portfolio object.
}
\examples{

n<-3e2 # number of realizations
p<-.5*n # number of assets
b<-rep(1/p,p)

# Assets with a diagonal covariance matrix
x <- matrix(data = rnorm(n*p), nrow = p, ncol = n)

test <- new_GMV_portfolio_weights_BDPS19(x=x, b=b, alph=0.05)
str(test)

# Assets with a non-diagonal covariance matrix
Mtrx <- RandCovMtrx(n=n, p=p, q=20.55)
x <- t(MASS::mvrnorm(n=n , mu=rep(0,p), Sigma=Mtrx))

test <- new_GMV_portfolio_weights_BDPS19(x=x, b=b, alph=0.05)
str(test)

}
\references{
\insertRef{BDPS2019}{hdsp}

\insertRef{BPS2018}{hdsp}
}
