% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDOPS_2020.R
\name{Sigma_sample_estimator}
\alias{Sigma_sample_estimator}
\title{Simple covariance estimator}
\usage{
Sigma_sample_estimator(x)
}
\arguments{
\item{x}{numeric matrix in which columns are independent realizations of asset returns}
}
\description{
Some description
}
\examples{
p<-5 # number of assets
n<-1e1 # number of realizations

x <-matrix(data = rnorm(n*p), nrow = p, ncol = n)
Sigma_sample_estimator(x)
}
