% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_custom_portfol.R
\name{validate_ExUtil_portfolio}
\alias{validate_ExUtil_portfolio}
\title{A validator for ExUtil_portfolio}
\usage{
validate_ExUtil_portfolio(x)
}
\arguments{
\item{x}{Object of class ExUtil_portfolio.}
}
\description{
A validator for ExUtil_portfolio
}
